+++ packages/orq-rc/src/models/operations/createbudget.ts
@@ -658,7 +658,7 @@ export const CreateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-02T12:46:46.580Z",
+    "2025-10-02T13:55:24.060Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -693,7 +693,7 @@ export const CreateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => Consumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-02T12:46:46.580Z"))
+  updated: z.date().default(() => new Date("2025-10-02T13:55:24.060Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {