+++ packages/orq-rc/src/models/operations/listagents.ts
@@ -1075,7 +1075,7 @@ export const ListAgentsKnowledgeBases$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01K6JFC6M7BBTDMVV5H8XPRX5D"),
+  id: z.string().default("01K6JK9V1ECE9ZM1QAXW824DDZ"),
   knowledge_id: z.string(),
   configuration: z.union([
     z.lazy(() =>
@@ -1106,7 +1106,7 @@ export const ListAgentsKnowledgeBases$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   ListAgentsKnowledgeBases
 > = z.object({
-  id: z.string().default("01K6JFC6M7BBTDMVV5H8XPRX5D"),
+  id: z.string().default("01K6JK9V1ECE9ZM1QAXW824DDZ"),
   knowledgeId: z.string(),
   configuration: z.union([
     z.lazy(() =>